requirements.txt
colorama==0.4.6
python-dotenv==1.0.1
requests==2.32.3
tabulate==0.9.0
yarl==1.9.4
